+++ crates/uv-git-types/src/oid.rs
