+++ crates/uv-git/src/resolver.rs
@@ -3,16 +3,18 @@ use std::path::PathBuf;
 use std::str::FromStr;
 use std::sync::Arc;
 
-use tracing::debug;
-
-use crate::{Fetch, GitHubRepository, GitOid, GitReference, GitSource, GitUrl, Reporter};
 use dashmap::mapref::one::Ref;
 use dashmap::DashMap;
 use fs_err::tokio as fs;
 use reqwest_middleware::ClientWithMiddleware;
+use tracing::debug;
+
 use uv_cache_key::{cache_digest, RepositoryUrl};
 use uv_fs::LockedFile;
+use uv_git_types::{GitHubRepository, GitOid, GitReference, GitUrl};
 use uv_version::version;
+
+use crate::{Fetch, GitSource, Reporter};
 
 #[derive(Debug, thiserror::Error)]
 pub enum GitResolverError {
