+++ crates/uv-git/src/source.rs
@@ -12,9 +12,10 @@ use tracing::{debug, instrument};
 use url::Url;
 
 use uv_cache_key::{cache_digest, RepositoryUrl};
+use uv_git_types::{GitOid, GitUrl};
 
 use crate::git::GitRemote;
-use crate::{GitOid, GitUrl, GIT_STORE};
+use crate::GIT_STORE;
 
 /// A remote Git source that can be checked out locally.
 pub struct GitSource {
@@ -65,10 +66,10 @@ impl GitSource {
     }
 
     /// Fetch the underlying Git repository at the given revision.
-    #[instrument(skip(self), fields(repository = %self.git.repository, rev = ?self.git.precise))]
+    #[instrument(skip(self), fields(repository = %self.git.repository(), rev = ?self.git.precise()))]
     pub fn fetch(self) -> Result<Fetch> {
         // Compute the canonical URL for the repository.
-        let canonical = RepositoryUrl::new(&self.git.repository);
+        let canonical = RepositoryUrl::new(self.git.repository());
 
         // The path to the repo, within the Git database.
         let ident = cache_digest(&canonical);
@@ -76,17 +77,17 @@ impl GitSource {
 
         // Authenticate the URL, if necessary.
         let remote = if let Some(credentials) = GIT_STORE.get(&canonical) {
-            Cow::Owned(credentials.apply(self.git.repository.clone()))
+            Cow::Owned(credentials.apply(self.git.repository().clone()))
         } else {
-            Cow::Borrowed(&self.git.repository)
+            Cow::Borrowed(self.git.repository())
         };
 
         let remote = GitRemote::new(&remote);
-        let (db, actual_rev, task) = match (self.git.precise, remote.db_at(&db_path).ok()) {
+        let (db, actual_rev, task) = match (self.git.precise(), remote.db_at(&db_path).ok()) {
             // If we have a locked revision, and we have a preexisting database
             // which has that revision, then no update needs to happen.
             (Some(rev), Some(db)) if db.contains(rev) => {
-                debug!("Using existing Git source `{}`", self.git.repository);
+                debug!("Using existing Git source `{}`", self.git.repository());
                 (db, rev, None)
             }
 
@@ -95,17 +96,17 @@ impl GitSource {
             // situation that we have a locked revision but the database
             // doesn't have it.
             (locked_rev, db) => {
-                debug!("Updating Git source `{}`", self.git.repository);
+                debug!("Updating Git source `{}`", self.git.repository());
 
                 // Report the checkout operation to the reporter.
                 let task = self.reporter.as_ref().map(|reporter| {
-                    reporter.on_checkout_start(remote.url(), self.git.reference.as_rev())
+                    reporter.on_checkout_start(remote.url(), self.git.reference().as_rev())
                 });
 
                 let (db, actual_rev) = remote.checkout(
                     &db_path,
                     db,
-                    &self.git.reference,
+                    self.git.reference(),
                     locked_rev.map(GitOid::from),
                     &self.client,
                     self.disable_ssl,
