+++ crates/uv-pypi-types/Cargo.toml
@@ -18,7 +18,7 @@ workspace = true
 [dependencies]
 uv-distribution-filename = { workspace = true }
 uv-fs = { workspace = true, features = ["serde"] }
-uv-git = { workspace = true }
+uv-git-types = { workspace = true }
 uv-normalize = { workspace = true }
 uv-pep440 = { workspace = true }
 uv-pep508 = { workspace = true }
